+++ RichMessageKit/Models/SuggestedReplyMessage.swift
@@ -8,12 +8,30 @@
 import Foundation
 
 public struct SuggestedReplyMessage {
-    /// Title to be displayed in suggested replies
-    public var title: [String]
 
-    /// Reply that should be given when title is clicked
-    /// If nil, then title will be used as reply
-    public var reply: [String?]
+    public enum SuggestionType {
+        case link
+        case normal
+    }
+
+    public struct Suggestion {
+        let title: String
+        let type: SuggestionType
+        /// Reply that should be given when title is clicked
+        /// If nil, then title will be used as reply
+        public var reply: String?
+
+        public init(title: String, reply: String? = nil, type: SuggestionType = .normal) {
+            self.title = title
+            self.reply = reply
+            self.type = type
+        }
+    }
+
+    /// Title to be displayed in the view.
+    /// Dictionary of name and type.
+    public var suggestion: [Suggestion]
+
 
     public var message: Message
 }