+++ Sources/Controllers/ALKConversationViewController.swift
@@ -642,6 +642,8 @@ open class ALKConversationViewController: ALKBaseViewController, Localizable {
         tableView.register(ReceivedImageMessageCell.self)
         tableView.register(ReceivedFAQMessageCell.self)
         tableView.register(SentFAQMessageCell.self)
+        tableView.register(SentButtonsCell.self)
+        tableView.register(ReceivedButtonsCell.self)
     }
 
     private func prepareMoreBar() {
@@ -1133,6 +1135,32 @@ open class ALKConversationViewController: ALKBaseViewController, Localizable {
         sendQuickReply(msg, metadata: metadata)
     }
 
+    func richButtonSelected(index: Int,
+                            title: String,
+                            message: ALKMessageViewModel,
+                            isButtonClickDisabled: Bool) {
+        guard
+            !isButtonClickDisabled,
+            let payload = message.payloadFromMetadata()?[index],
+            let action = payload["action"] as? [String: Any],
+            let type = action["type"] as? String
+        else {
+            return
+        }
+        switch type {
+        case "link":
+            linkButtonSelected(action)
+        case "submit":
+            let ackMessage = action["message"] as? String ?? title
+            submitButtonSelected(metadata: action, text: ackMessage)
+        case "quickReply":
+            let ackMessage = action["message"] as? String ?? title
+            sendQuickReply(ackMessage, metadata: payload["replyMetadata"] as? [String: Any])
+        default:
+            print("Do nothing")
+        }
+    }
+
     func messageButtonSelected(
         index: Int,
         title: String,
@@ -1304,6 +1332,7 @@ open class ALKConversationViewController: ALKBaseViewController, Localizable {
 
     private func sendQuickReply(_ text: String, metadata: [String: Any]?) {
         var customMetadata = metadata ?? [String: Any]()
+
         guard let messageMetadata = configuration.messageMetadata as? [String: Any] else {
             viewModel.send(message: text, metadata: customMetadata)
             return
@@ -1312,11 +1341,11 @@ open class ALKConversationViewController: ALKBaseViewController, Localizable {
         viewModel.send(message: text, metadata: customMetadata)
     }
 
-    private func postRequestUsing(url: URL, param: String) -> URLRequest? {
+    private func postRequestUsing(url: URL, param: [String: Any]) -> URLRequest? {
         var request = URLRequest(url: url)
         request.timeoutInterval = 600
         request.httpMethod = "POST"
-        guard let data = param.data(using: .utf8) else { return nil }
+        guard let data = ALUtilityClass.generateJsonString(from: param)?.data(using: .utf8) else { return nil }
         request.httpBody = data
         let contentLength = String(format: "%lu", UInt(data.count))
         request.setValue(contentLength, forHTTPHeaderField: "Content-Length")
@@ -1378,7 +1407,7 @@ open class ALKConversationViewController: ALKBaseViewController, Localizable {
 
     private func submitButtonSelected(metadata: [String: Any], text: String) {
         guard
-            let formData = metadata["formData"] as? String,
+            let formData = metadata["formData"] as? [String: Any],
             let urlString = metadata["formAction"] as? String,
             let url = URL(string: urlString),
             var request = postRequestUsing(url: url, param: formData)